move to general setting
