+++ apps/obsidian/src/components/Settings.tsx
@@ -1,5 +1,5 @@
-import { StrictMode, useState, useRef, useEffect } from "react";
-import { App, PluginSettingTab, setIcon } from "obsidian";
+import { StrictMode, useState } from "react";
+import { App, PluginSettingTab } from "obsidian";
 import type DiscourseGraphPlugin from "~/index";
 import { Root, createRoot } from "react-dom/client";
 import { ContextProvider } from "./AppContext";
@@ -7,88 +7,7 @@ import RelationshipTypeSettings from "./RelationshipTypeSettings";
 import RelationshipSettings from "./RelationshipSettings";
 import NodeTypeSettings from "./NodeTypeSettings";
 import GeneralSettings from "./GeneralSettings";
-import { PluginProvider, usePlugin } from "./PluginContext";
-import { SLACK_LOGO, WHITE_LOGO_SVG } from "~/icons";
-
-const DOCS_URL = "https://discoursegraphs.com/docs/obsidian";
-const COMMUNITY_URL =
-  "https://join.slack.com/t/discoursegraphs/shared_invite/zt-37xklatti-cpEjgPQC0YyKYQWPNgAkEg";
-
-const InfoSection = () => {
-  const plugin = usePlugin();
-  const logoRef = useRef<HTMLDivElement>(null);
-  const communityIconRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (logoRef.current) {
-      logoRef.current.innerHTML = WHITE_LOGO_SVG;
-    }
-    if (communityIconRef.current) {
-      communityIconRef.current.innerHTML = SLACK_LOGO;
-    }
-  }, []);
-
-  return (
-    <div className="flex justify-center">
-      <div
-        className="flex w-48 flex-col items-center rounded-lg p-3"
-        style={{ background: "var(--tag-background)" }}
-      >
-        <div
-          ref={logoRef}
-          className="h-12 w-12"
-          style={{ color: "var(--interactive-accent)" }}
-        />
-        <div
-          className="font-semibold"
-          style={{ color: "var(--interactive-accent)" }}
-        >
-          Discourse Graphs
-        </div>
-
-        <a
-          href={COMMUNITY_URL}
-          className="flex items-center gap-1 text-sm no-underline hover:opacity-80"
-          style={{ color: "var(--interactive-accent)" }}
-          target="_blank"
-          rel="noopener noreferrer"
-          aria-label="Community"
-        >
-          <div ref={communityIconRef} className="icon" />
-          <span>Community</span>
-          <span
-            className="icon"
-            ref={(el) => (el && setIcon(el, "arrow-up-right")) || undefined}
-          />
-        </a>
-        <a
-          href={DOCS_URL}
-          className="flex items-center gap-1 text-sm no-underline hover:opacity-80"
-          style={{ color: "var(--interactive-accent)" }}
-          target="_blank"
-          rel="noopener noreferrer"
-          aria-label="Docs"
-        >
-          <div
-            className="icon"
-            ref={(el) => (el && setIcon(el, "book")) || undefined}
-          />
-          <span>Docs</span>
-          <span
-            className="icon"
-            ref={(el) => (el && setIcon(el, "arrow-up-right")) || undefined}
-          />
-        </a>
-        <span
-          className="text-muted text-xs"
-          style={{ color: "var(--interactive-accent)" }}
-        >
-          {plugin.manifest.version}
-        </span>
-      </div>
-    </div>
-  );
-};
+import { PluginProvider } from "./PluginContext";
 
 const Settings = () => {
   const [activeTab, setActiveTab] = useState("general");
@@ -144,8 +63,6 @@ const Settings = () => {
       {activeTab === "relationTypes" && <RelationshipTypeSettings />}
       {activeTab === "relations" && <RelationshipSettings />}
       {activeTab === "frontmatter" && <GeneralSettings />}
-
-      <InfoSection />
     </div>
   );
 };
